All latest changes for launch

+++ scripts/populate.py
@@ -25,24 +25,26 @@ def user_json(row):
         "password": row[1],
         "firstName": row[2],
         "lastName": row[3],
-        "twitter": row[4],
-        "flickr": row[5],
+        "phone": row[4],
+        "homepage": row[5],
         "instagram": row[6],
-        "role": row[7],
-        "bio": row[8],
-        "tags": row[9]
+        "youtube": row[7],
+        "role": row[8],
+        "showChat": row[9] == 'TRUE',
+        "bio": row[10],
+        "tags": row[11]
     }
 
 def host_json(id, user):
     return {
-        # "id": int(id),
         "tags": user["tags"],
         "firstName": user["firstName"],
         "lastName": user["lastName"],
-        "bio": user["bio"],        
-        "twitter": user["twitter"],
-        "flickr": user["flickr"],
+        "bio": user["bio"],
+        "phone": user["phone"],
+        "homepage": user["homepage"],
         "instagram": user["instagram"],
+        "youtube": user["youtube"],
     }
 
 def place_json(row, host):
@@ -51,15 +53,26 @@ def place_json(row, host):
         "name": row[1],
         "address": row[2],
         "description": row[3],
-        "tags": row[4]
+        "homepage": row[4],
+        "instagram": row[5],
+        "facebook": row[6],
+        "matterport_link": row[7],
+        "tags": row[8]
     }
 
 def artwork_json(row, artist):
+    year = None
+    if row[5]:
+        year = int(row[5])
     return {
         "artist": artist,
         "name": row[1],
         "description": row[2],
-        "tags": row[4]
+        "medium": row[3],
+        "size": row[4],
+        "year": year,
+        "link": row[6],
+        "tags": row[7]
     }
 
 def event_json(place, artists, artworks, row):
@@ -69,9 +82,11 @@ def event_json(place, artists, artworks, row):
         "artworks": artworks,
         "name": row[3],
         "description": row[4],
-        "startTime": row[5],
-        "endTime": row[6],
-        "tags": row[7]
+        "link": row[5],
+        "hubs_link": row[6],
+        "startTime": row[7],
+        "endTime": row[8],
+        "tags": row[9]
     }
 
 def upload_image(bdir, fn, rkind, rid, user):
@@ -141,47 +156,51 @@ def upload_images_from_list(bdir, fnlist, rkind, rid, user):
 parser = argparse.ArgumentParser(description='Upload dummy data to OASIS server.')
 parser.add_argument('-u', '--url', action='store', default='http://127.0.0.1:5000', help='set server url')
 parser.add_argument('-a', '--admin', action='store', default='Admin Oasis', help='admin username')
+parser.add_argument('-f', '--folder', action='store', default='dummy_data', help='set base data folder')
+parser.add_argument('-s', '--step', action='store', default=0, type=int, help='set starting step of db population (0-8')
 args = parser.parse_args()
 
 server_url = args.url
 admin_name = args.admin
 
-data_dir = join(sys.path[0], "dummy_data")
+data_dir = join(sys.path[0], args.folder)
+step = args.step
 
 mimetypes.init()
 
-print("Populating users...")
+print("1 - Populating users...")
 in_csv = join(data_dir, "user_list.csv")
 reader = csv.reader(open(in_csv, 'r'), dialect='excel')
 header = next(reader)
 user_extra = {}
 for row in reader:
-    email = row[0]
-    password = row[1]
-    user_extra[row[2] + ' ' + row[3]] = {'email': email, 'password':password}
+    email = row[0].strip()
+    password = row[1].strip()
+    user_extra[(row[2] + ' ' + row[3]).strip()] = {'email': email, 'password': password}
     raw_user_data = user_json(row)
 
-    print("Creating user", row[2], row[3], "...")
-    user_data = make_data_request(raw_user_data)
-    r = requests.post(server_url + '/api/user/', data=user_data)
-    if r.status_code == 400:
-        print("  User already exists")
-        continue
-        
-    if r.status_code != 201:                
-        raise Exception(r.status_code, r.content)
+    if step <= 1:
+        print("Creating user", row[2], row[3], "...")
+        user_data = make_data_request(raw_user_data)
+        r = requests.post(server_url + '/api/user/', data=user_data)
+        if r.status_code == 400:
+            print("  User already exists")
+            continue
+            
+        if r.status_code != 201:                
+            raise Exception(r.status_code, r.content)
 
-    uid = r.json()["id"]
-    print("  Created user with id", uid)
+        uid = r.json()["id"]
+        print("  Created user with id", uid)
 
-# Uploading user images
+print("2 - Uploading user images...")
 user_dict = {}
 r = requests.get(server_url + '/api/user/')
 if r.status_code != 200:
     raise Exception(r.status_code)
 users = r.json()['users']
 for user in users:
-    fullName = user['firstName'] + ' ' + user['lastName']
+    fullName = (user['firstName'] + ' ' + user['lastName']).strip()
     if not fullName == admin_name:
         user['email'] = user_extra[fullName]['email']
         user['password'] = user_extra[fullName]['password']
@@ -193,101 +212,117 @@ for user in users:
         base_path = data_dir + "/images/users/hosts/" + user['email']
     elif role == 3:
         base_path = data_dir + "/images/users/artists/" + user['email']
-    print("Uploading images for user", user["firstName"], user["lastName"])
-    upload_images_from_folder(base_path, "user", uid, user)
+    if step <= 2:
+        print("Uploading images for user", user["firstName"], user["lastName"])
+        upload_images_from_folder(base_path, "user", uid, user)
 
-print("Populating artworks...")
+print("3 - Populating artworks...")
 in_csv = join(data_dir, "artwork_list.csv")
 reader = csv.reader(open(in_csv, 'r'), dialect='excel')
-header = next(reader)    
+header = next(reader)
 artwork_images = {}
 for row in reader:
-    print("Creating artwork", row[1], "...")
-    print("  Logging user", row[0])    
     user = user_dict[row[0]]
+    if step <= 3:
+        print("Creating artwork", row[1], "...")
 
-    # First the host user needs to login so we have the token to use in place creation
-    user_data = make_data_request({'email': user['email'], 'password': user['password']})
-    r = requests.post(server_url + '/api/login/', data=user_data)
-    if r.status_code != 200:
-        raise Exception(r.status_code, r.content)
-    host_token = r.json()['token']
-    host_header = auth_header(host_token)
+        # First the host user needs to login so we have the token to use in place creation
+        print("  Logging user", row[0])        
+        user_data = make_data_request({'email': user['email'], 'password': user['password']})
+        r = requests.post(server_url + '/api/login/', data=user_data)
+        if r.status_code != 200:
+            raise Exception(r.status_code, r.content)
+        host_token = r.json()['token']
+        host_header = auth_header(host_token)
 
-    artist = {"id": user_dict[row[0].strip()]['id']}
+        artist = {"id": user_dict[row[0].strip()]['id']}
 
-    raw_artwork_data = artwork_json(row, artist)
-    p = make_data_request(raw_artwork_data)
-    r = requests.post(server_url + '/api/artwork/', data=p, headers=host_header)
+        raw_artwork_data = artwork_json(row, artist)
+        p = make_data_request(raw_artwork_data)
+        r = requests.post(server_url + '/api/artwork/', data=p, headers=host_header)
 
-    if r.status_code != 201:
-        raise Exception(r.status_code, r.content)
+        if r.status_code != 201:
+            raise Exception(r.status_code, r.content)
 
-    pid = r.json()["id"]
-    artwork_images[pid] = row[3].split(";")
+        pid = r.json()["id"]
+        artwork_images[pid] = row[8].split(";")
 
-    print("  Created artwork with id", pid)
+        print("  Created artwork with id", pid)
 
-    # Logout
-    r = requests.delete(server_url + '/api/login/', headers=host_header)
-    if r.status_code != 200:
-        raise Exception(r.status_code, r.content)    
-        
-    print("  Logged out succesfully")        
+        # Logout
+        r = requests.delete(server_url + '/api/login/', headers=host_header)
+        if r.status_code != 200:
+            raise Exception(r.status_code, r.content)    
+            
+        print("  Logged out succesfully")
 
-# Uploading artwork images
+print("4 - Uploading artwork images...")
 artwork_dict = {}
 r = requests.get(server_url + '/api/artwork/')
 if r.status_code != 200:
     raise Exception(r.status_code)
 artworks = r.json()['artworks']
+counts = {}
 for artwork in artworks:    
     pid = artwork['id']
-    name = artwork['name']
+    name = artwork['name'].strip()
+    if not name: name = 'Untitled'
     artist = artwork['artist']
-    artwork_dict[name] = artwork
-    user = user_dict[artist['firstName'] + ' ' + artist['lastName']]
-    base_path = data_dir + "/images/artworks/" + user["email"]
-    images = artwork_images[pid]
-    print("Uploading images for artwork", artwork["name"])
-    upload_images_from_list(base_path, images, "artwork", pid, user)
+    fname = (artist['firstName'] + ' ' + artist['lastName']).strip()
+    if fname in counts:
+        count = counts[fname]
+        count += 1
+    else:
+        count = 1
+    counts[fname] = count
+    key = name
+    if name == 'Untitled': key = fname + ':' + str(count)
+    artwork_dict[key] = artwork
+    if step <= 4:
+        user = user_dict[fname]
+        base_path = data_dir + "/images/artworks/" + user["email"]
+        images = artwork_images[pid]
+        print("Uploading images for artwork", artwork["name"])
+        upload_images_from_list(base_path, images, "artwork", pid, user)
 
-print("Populating places...")
+#print(artwork_dict)
+
+print("5 - Populating places...")
 in_csv = join(data_dir, "place_list.csv")
 reader = csv.reader(open(in_csv, 'r'), dialect='excel')
 header = next(reader)
 for row in reader:
-    print("Creating place", row[1], "...")
-    print("  Logging user", row[0])
-
     user = user_dict[row[0]]
+    if step <= 5:
+        print("Creating place", row[1], "...")    
 
-    # First the host user needs to login so we have the token to use in place creation
-    user_data = make_data_request({'email': user['email'], 'password': user['password']})
-    r = requests.post(server_url + '/api/login/', data=user_data)
-    if r.status_code != 200:
-        raise Exception(r.status_code, r.content)
-    host_token = r.json()['token']
-    host_header = auth_header(host_token)
+        # First the host user needs to login so we have the token to use in place creation
+        print("  Logging user", row[0])    
+        user_data = make_data_request({'email': user['email'], 'password': user['password']})
+        r = requests.post(server_url + '/api/login/', data=user_data)
+        if r.status_code != 200:
+            raise Exception(r.status_code, r.content)
+        host_token = r.json()['token']
+        host_header = auth_header(host_token)
 
-    raw_place_data = place_json(row, host_json(row[0], user))
-    p = make_data_request(raw_place_data)
-    r = requests.post(server_url + '/api/place/', data=p, headers=host_header)
+        raw_place_data = place_json(row, host_json(row[0], user))
+        p = make_data_request(raw_place_data)
+        r = requests.post(server_url + '/api/place/', data=p, headers=host_header)
 
-    if r.status_code != 201:
-        raise Exception(r.status_code, r.content)
+        if r.status_code != 201:
+            raise Exception(r.status_code, r.content)
 
-    pid = r.json()["id"]
-    print("  Created place with id", pid)
+        pid = r.json()["id"]
+        print("  Created place with id", pid)
 
-    # Logout
-    r = requests.delete(server_url + '/api/login/', headers=host_header)
-    if r.status_code != 200:
-        raise Exception(r.status_code, r.content)    
-        
-    print("  Logged out succesfully")
+        # Logout
+        r = requests.delete(server_url + '/api/login/', headers=host_header)
+        if r.status_code != 200:
+            raise Exception(r.status_code, r.content)    
+            
+        print("  Logged out succesfully")
 
-# Uploading place images
+print("6 - Uploading place images...")
 place_dict = {}
 r = requests.get(server_url + '/api/place/')
 if r.status_code != 200:
@@ -295,14 +330,15 @@ if r.status_code != 200:
 places = r.json()['places']
 for place in places:
     place_dict[place['name']] = place
-    pid = place['id']
-    host = place['host']
-    user = user_dict[host['firstName'] + ' ' + host['lastName']]
-    base_path = data_dir + "/images/places/" + place["name"]
-    print("Uploading images for place", place["name"])
-    upload_images_from_folder(base_path, "place", pid, user)
+    if step <= 6:
+        pid = place['id']
+        host = place['host']
+        user = user_dict[(host['firstName'] + ' ' + host['lastName']).strip()]
+        base_path = data_dir + "/images/places/" + place["name"]
+        print("Uploading images for place", place["name"])
+        upload_images_from_folder(base_path, "place", pid, user)
 
-print("Populating events...")
+print("7 - Populating events...")
 in_csv = join(data_dir, "event_list.csv")
 reader = csv.reader(open(in_csv, 'r'), dialect='excel')
 header = next(reader)
@@ -312,10 +348,10 @@ rows = []
 first_date = None
 date_format = '%Y-%m-%dT%H:%M:%S'
 for row in reader:
-    start_date = datetime.strptime(row[5], date_format)
-    end_date = datetime.strptime(row[6], date_format)
-    row[5] = start_date
-    row[6] = end_date
+    start_date = datetime.strptime(row[7], date_format)
+    end_date = datetime.strptime(row[8], date_format)
+    row[7] = start_date
+    row[8] = end_date
     if not first_date: first_date = start_date
     if start_date < first_date:
         first_date = start_date
@@ -325,53 +361,59 @@ NOW = datetime.now()
 diff = NOW - first_date
 
 for row in rows:
-    event_extra[row[3]] = {'image': row[8]}
-    print("Creating event", row[3], "...")
+    event_extra[row[3]] = {'image': row[10]}
 
     # Normalizing dates using today as reference
-    start_date = row[5] + diff
-    end_date = row[6] + diff
-    row[5] = start_date.strftime(date_format)
-    row[6] = end_date.strftime(date_format)
+    start_date = row[7] + diff
+    end_date = row[8] + diff
+    row[7] = start_date.strftime(date_format)
+    row[8] = end_date.strftime(date_format)
 
     place = {"id": place_dict[row[0]]['id']}
-    artists = [{"id":user_dict[name.strip()]['id']} for name in row[1].split(';')]
-    artworks = [{"id":artwork_dict[name.strip()]['id']} for name in row[2].split(';')]
-
+    if row[1]:
+        artists = [{"id":user_dict[name.strip()]['id']} for name in row[1].split(';')]
+    else:
+        artists = []
+    if row[2]:        
+        artworks = [{"id":artwork_dict[name.strip()]['id']} for name in row[2].split(';')]
+    else:
+        artworks = []
     host = place_dict[row[0]]['host']
-    hostFullName = host['firstName'] + ' ' + host['lastName']
+    hostFullName = (host['firstName'] + ' ' + host['lastName']).strip()
     hostEmail = user_dict[hostFullName]['email']
     hostPassword = user_dict[hostFullName]['password']
 
-    print("  Logging host", hostFullName)
-        
-    # First the host user needs to login so we have the token to use in place creation
-    user_data = make_data_request({'email': hostEmail, 'password': hostPassword})
-    r = requests.post(server_url + '/api/login/', data=user_data)
-    if r.status_code != 200:
-        raise Exception(r.status_code, r.content)
-    host_token = r.json()['token']
-    host_header = auth_header(host_token)   
+    if step <= 7:
+        print("Creating event", row[3], "...")
 
-    raw_event_data = event_json(place, artists, artworks, row)
-    user_data = make_data_request(raw_event_data)
+        # First the host user needs to login so we have the token to use in place creation
+        print("  Logging host", hostFullName)        
+        user_data = make_data_request({'email': hostEmail, 'password': hostPassword})
+        r = requests.post(server_url + '/api/login/', data=user_data)
+        if r.status_code != 200:
+            raise Exception(r.status_code, r.content)
+        host_token = r.json()['token']
+        host_header = auth_header(host_token)   
 
-    r = requests.post(server_url + '/api/event/', data=user_data, headers=host_header)
+        raw_event_data = event_json(place, artists, artworks, row)
+        user_data = make_data_request(raw_event_data)
 
-    if r.status_code != 201:
-        raise Exception(r.status_code, r.content)
+        r = requests.post(server_url + '/api/event/', data=user_data, headers=host_header)
 
-    eid = r.json()["id"]
-    print("  Created event with id", eid)
+        if r.status_code != 201:
+            raise Exception(r.status_code, r.content)
 
-    # Logout
-    r = requests.delete(server_url + '/api/login/', headers=host_header)
-    if r.status_code != 200:
-        raise Exception(r.status_code, r.content)    
-        
-    print("  Logged out succesfully")
+        eid = r.json()["id"]
+        print("  Created event with id", eid)
 
-# Uploading event images
+        # Logout
+        r = requests.delete(server_url + '/api/login/', headers=host_header)
+        if r.status_code != 200:
+            raise Exception(r.status_code, r.content)    
+            
+        print("  Logged out succesfully")
+
+print("8 - Uploading event images...")
 events_dict = {}
 r = requests.get(server_url + '/api/event/')
 if r.status_code != 200:
@@ -379,9 +421,10 @@ if r.status_code != 200:
 events = r.json()['events']
 for event in events:
     if not event['name'] in event_extra: continue
-    eid = event['id']
-    host = event['place']['host']
-    user = user_dict[host['firstName'] + ' ' + host['lastName']]
-    fn = event_extra[event['name']]['image']
-    print("Uploading images for event", event["name"])
-    upload_image(data_dir + "/images/events", fn, "event", eid, user)
+    if step <= 8:
+        eid = event['id']
+        host = event['place']['host']
+        user = user_dict[(host['firstName'] + ' ' + host['lastName']).strip()]
+        fn = event_extra[event['name']]['image']
+        print("Uploading images for event", event["name"])
+        upload_image(data_dir + "/images/events", fn, "event", eid, user)
